Fix some field sizes in P2Pv2 header.
--HG--
branch : soc.2010.msn-tlc
extra : convert_revision : a52197a5d3f84d3783b129aa8faee5e7ef827e5b

+++ libpurple/protocols/msn/p2p.h
@@ -21,8 +21,8 @@ typedef struct {
 typedef struct {
 	guint8  header_len;
 	guint8  opcode;
-	guint32 message_len;
-	guint64 base_id;
+	guint16 message_len;
+	guint32 base_id;
 } MsnP2Pv2Header;
 #pragma pack(pop)
